requirements.txt
influxdb
requests
opentsdb-py
psycopg2-binary
httplib2